requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file /var/app/src/api/requirements.txt /var/app/src/api/requirements.in
#
amqp==2.5.2               # via kombu
attrs==19.3.0             # via jsonschema, pytest
autopep8==1.4.4           # via django-silk
aws-sam-translator==1.19.0  # via cfn-lint
billiard==3.5.0.5         # via celery
certifi==2019.11.28       # via requests
cffi==1.13.2              # via cryptography
chardet==3.0.4            # via requests
coverage==4.5.1
cryptography==2.8         # via moto, sshpubkeys
des==1.0.5
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-filter==2.1.0
django-queryinspect==0.1.0
django-redis==4.5.0
django-silk==3.0.4
django==2.2
djangorestframework==3.9.2
docutils==0.15.2          # via botocore
ecdsa==0.14.1             # via python-jose, sshpubkeys
factory-boy==2.12.0
faker==3.0.0              # via factory-boy
flake8==3.6.0
future==0.18.2            # via aws-xray-sdk, python-jose
gprof2dot==2019.11.30     # via django-silk
gunicorn==19.7.1
idna==2.8                 # via moto, requests
importlib-metadata==1.2.0  # via jsonschema, kombu, pluggy, pytest
importlib-resources==1.0.2  # via cfn-lint
jinja2==2.10.3            # via django-silk, moto
jmespath==0.9.4           # via boto3, botocore
jsonpatch==1.24           # via cfn-lint
jsonpickle==1.2           # via aws-xray-sdk
jsonpointer==2.0          # via jsonpatch
jsonschema==3.2.0         # via aws-sam-translator, cfn-lint
kombu==4.6.6              # via celery
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==8.0.1     # via pytest, zipp
packaging==19.2           # via pytest
pathlib2==2.3.5           # via pytest
pillow==6.2.1
pluggy==0.13.1            # via pytest
psycopg2-binary==2.7.7
py==1.8.0                 # via pytest
pyasn1==0.4.8             # via rsa
pycodestyle==2.4.0        # via autopep8, flake8
pycparser==2.19           # via cffi
pyflakes==2.0.0           # via flake8
pygments==2.5.2           # via django-silk
pyparsing==2.4.5          # via packaging
pyrsistent==0.15.6        # via jsonschema
pytest-django==3.4.8
pytest==5.3.1             # via pytest-django
python-dateutil==2.8.0    # via botocore, celery-redbeat, django-silk, faker, moto
pytz==2019.3              # via celery, django, django-silk, moto
pyyaml==5.2               # via cfn-lint, moto
raven==6.10.0
redis==2.10.5
requests==2.22.0
rsa==4.0                  # via python-jose
s3transfer==0.2.1         # via boto3
six==1.13.0               # via aws-sam-translator, cfn-lint, cryptography, docker, ecdsa, faker, jsonschema, mock, moto, packaging, pathlib2, pyrsistent, python-dateutil, python-jose, responses, websocket-client
sqlparse==0.3.0           # via django, django-silk
text-unidecode==1.3       # via faker
urllib3==1.25.7           # via botocore, requests
vine==1.3.0               # via amqp
virtualenv==13.1.2
wcwidth==0.1.7            # via pytest
websocket-client==0.56.0  # via docker
wrapt==1.11.2             # via aws-xray-sdk
zipp==0.6.0               # via importlib-metadata
django-cors-headers==3.2.1
django-rest-swagger==2.2.0